cctu: A package for producing analysis reports from clinical trials

The package `cctu` is designed to help run the analysis of clinical trials which typically produce an output of a large word document with one page per table or figure following on from a Statistical Analysis Plan, and requires substantial audit trails. Automating the trivialities of getting tables into word from R, and capturing the meta data and audit trials is what the package achieves.

Details

See the vignette vignette("analysis-template", package="cctu") for an overview and starting point.

Key Functions

meta_table_example

create_popn_envir

attach_pop

sumby

write_table

write_plot

write_docx

cttab

apply_macro_dict

regression_table

See also

Author

Maintainer: Simon Bond simon.bond7@nhs.net

Authors:

Alimu Dayimu