Render categorical values for table output.

Called from cttab by default to render categorical (i.e. factor, character or logical) values for displaying in the table.

Usage

render_cat(x, ...)

Arguments

x

A vector of type factor, character or logical.

...

Further arguments, passed to cat_stat.

Value

A character vector. Each element is to be displayed in a separate cell in the table. The names of the vector are the labels to use in the table.

Details

This function was used by link{cttab} to render categorical variables. It essentially uses the values returned by cat_stat and put values to a vector. You can modified this to show any values you want, checkout the example below.

See also

Examples

y <- factor(sample(0:1, 99, replace = TRUE), labels = c("Female", "Male"))
y[1:10] <- NA
render_cat(y)
#>          Female            Male 
#> "36/89 (40.4%)" "53/89 (59.6%)"