Produces the final HTML, PDF or DOCX file

This is for generating other document format type with Quarto. The created docx and pdf documents may not have a nice output. Hence, the write_docx should be used for the docx or pdf by default. The PDF can be converted with Word. The rendering time for the the HTML is much faster than PDF. Tables will be formated with flextable and rendered with predifined Quarto templates using quarto_render.

Usage

write_quarto(
  report_title,
  author,
  meta_table = get_meta_table(),
  filename = file.path(getOption("cctu_output", default = "Output"), "Reports",
    "Report.html"),
  table_path = file.path(getOption("cctu_output", default = "Output"), "Core"),
  figure_path = file.path(getOption("cctu_output", default = "Output"), "Figures"),
  output_format = c("html", "pdf", "docx"),
  popn_labels = NULL,
  verbose = options()$verbose,
  quiet = TRUE,
  ...
)

Arguments

report_title

text string used to label the report title page

author

text string naming the author

meta_table

a data frame that contains meta information on tables (title, population, number). Defaults is get_meta_table()

filename

text string giving the filename/path to output the word document to

table_path

text string giving the filepath to the tables folder. This is used to directly open table files and copy the content.

figure_path

text string giving the filepath to the figures folder. This is used to create a reference within the word document. Hence it needs to be a relative (to the output report) path "../Figures". Likely to lead to empty figures in the report if it is changed.

output_format

Output format of the document, can be "html", "pdf" or "docx".

popn_labels

alternative text string giving labels used for the population - might want to include the population size... They must match correctly to unique(meta_table$population), excluding rows with a blank, or no, population given

verbose

logical to print information on changes to the global environment or external files. Defaults to options()$verbose.

quiet

Suppress warning and other messages from quarto, see quarto_render

...

Other parameters passed to quarto_render.

Value

This function is run for its side-effects: creates an HTML document.

Details

suggest that file.path is used to create non default file paths, to cope with OS vaguaries.